Move OmniAuth logic to a concern. Tweak logger levels, Add a provider_linked notification
