+++ app/views/application_mailer/provider_linked.html.haml
@@ -1,0 +1,14 @@
+%p.intro
+  Hello #{@user.name},
+
+%p.content
+  We're contacting you to let you know that your account at #{link_to _('City of Boston DND Warehouse'), root_url} has been signed into with #{@provider_name} for the first time.
+
+%p.content
+  You should now use #{@provider_name} to sign in, your original password has been disabled.
+
+%p.content
+  If you did expect this please contact your administrator.
+
+%p.closing
+  Thank you.
