+++ config/environments/development.rb
@@ -52,11 +52,13 @@ Rails.application.configure do
         port: smtp_port,
       }
     end
+    puts "Using stmp #{ENV['SMTP_SERVER']} for mail delivery"
   else
     # Don't care if the mailer can't send.
     config.action_mailer.raise_delivery_errors = false
 
     config.action_mailer.delivery_method = ENV.fetch("DEV_MAILER") { :file }.to_sym
+    puts "Using #{config.action_mailer.delivery_method} for mail delivery"
   end
 
   config.action_mailer.perform_caching = false
